Move migrated pool into separate package
