+++ internal/async/pool.go
@@ -1,5 +1,3 @@
-//go:build !migrated_fynedo
-
 package async
 
 import "sync"
